Added cleansed data from cmumps (NC) database to round out testing regime.
Some keys that appear in schema appear in NO document in the cmumps database. Commented code to that effect but still invented values in cmumps-patient7.jsonld.

Augmented notes.md to describe how to generate the set of *actual* keys that appear in all documents in a collection. This was lifted from the Interwebs.